Add support for float seconds in typesAndWaits() method (#765)
* Add support for float seconds in typesAndWaits() method

* Fix missing update in the tests folder

+++ src/Drivers/Tests/ProxyDriver.php
@@ -108,10 +108,10 @@ final class ProxyDriver implements DriverInterface
     /**
      * Send a typing indicator and wait for the given amount of seconds.
      * @param IncomingMessage $matchingMessage
-     * @param int $seconds
+     * @param float $seconds
      * @return mixed
      */
-    public function typesAndWaits(IncomingMessage $matchingMessage, int $seconds)
+    public function typesAndWaits(IncomingMessage $matchingMessage, float $seconds)
     {
         return self::instance()->typesAndWaits($matchingMessage, $seconds);
     }
