+++ tests/Fixtures/TestDriverWithSubDriver.php
@@ -144,10 +144,10 @@ class TestDriverWithSubDriver implements DriverInterface
     /**
      * Send a typing indicator and wait for the given amount of seconds.
      * @param IncomingMessage $matchingMessage
-     * @param int $seconds
+     * @param float $seconds
      * @return mixed
      */
-    public function typesAndWaits(IncomingMessage $matchingMessage, int $seconds)
+    public function typesAndWaits(IncomingMessage $matchingMessage, float $seconds)
     {
     }
 }
